Bluetooth: GATT: Fix error type
Callback errors should contain ATT error code not posix errors.


Former-commit-id: 6a4d39d5cde11e5c495a1cb6832ea116a1f3582c

+++ include/bluetooth/gatt.h
@@ -59,7 +59,7 @@ DECL|bt_gatt_include|struct|struct bt_gatt_include {
 DECL|bt_gatt_indicate_func_t|typedef|typedef void (*bt_gatt_indicate_func_t)(struct bt_conn *conn,
 DECL|bt_gatt_indicate_params|struct|struct bt_gatt_indicate_params {
 DECL|bt_gatt_notify_func_t|typedef|typedef uint8_t (*bt_gatt_notify_func_t)(struct bt_conn *conn,
-DECL|bt_gatt_read_func_t|typedef|typedef uint8_t (*bt_gatt_read_func_t)(struct bt_conn *conn, int err,
+DECL|bt_gatt_read_func_t|typedef|typedef uint8_t (*bt_gatt_read_func_t)(struct bt_conn *conn, uint8_t err,
 DECL|bt_gatt_read_params|struct|struct bt_gatt_read_params {
 DECL|bt_gatt_rsp_func_t|typedef|typedef void (*bt_gatt_rsp_func_t)(struct bt_conn *conn, uint8_t err);
 DECL|bt_gatt_service|struct|struct bt_gatt_service {
